Applying authorization.

+++ RacingLeagueManager/Pages/LeagueDriver/Index.cshtml.cs
@@ -3,21 +3,25 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
+using RacingLeagueManager.Authorization;
 using RacingLeagueManager.Data;
 using RacingLeagueManager.Data.Models;
+using RacingLeagueManager.Pages.Shared;
 
 namespace RacingLeagueManager.Pages.LeagueDriver
 {
-    public class IndexModel : PageModel
+    public class IndexModel : DI_BasePageModel
     {
-        private readonly RacingLeagueManager.Data.RacingLeagueManagerContext _context;
-
-        public IndexModel(RacingLeagueManager.Data.RacingLeagueManagerContext context)
+        public IndexModel(RacingLeagueManagerContext context,
+            IAuthorizationService authorizationService,
+            UserManager<Driver> userManager)
+        : base(context, authorizationService, userManager)
         {
-            _context = context;
         }
 
         //public IList<Data.Models.LeagueDriver> LeagueDriver { get;set; }
@@ -73,11 +77,19 @@ namespace RacingLeagueManager.Pages.LeagueDriver
                 NotFound();
             }
 
-            var leagueDriver = await _context.LeagueDriver.FirstOrDefaultAsync(l => l.LeagueId == leagueId && l.DriverId == driverId);
+            var leagueDriver = await _context.LeagueDriver.Include(l => l.League).FirstOrDefaultAsync(l => l.LeagueId == leagueId && l.DriverId == driverId);
 
             if(leagueDriver == null)
             {
                 NotFound();
+            }
+
+            var isAuthorized = await _authorizationService.AuthorizeAsync(
+                                                User, leagueDriver.League,
+                                                Operations.Update);
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
             }
 
             leagueDriver.Status = "Active";
@@ -94,11 +106,19 @@ namespace RacingLeagueManager.Pages.LeagueDriver
                 NotFound();
             }
 
-            var leagueDriver = await _context.LeagueDriver.FirstOrDefaultAsync(l => l.LeagueId == leagueId && l.DriverId == driverId);
+            var leagueDriver = await _context.LeagueDriver.Include(l => l.League).FirstOrDefaultAsync(l => l.LeagueId == leagueId && l.DriverId == driverId);
 
             if (leagueDriver == null)
             {
                 NotFound();
+            }
+
+            var isAuthorized = await _authorizationService.AuthorizeAsync(
+                                                User, leagueDriver.League,
+                                                Operations.Update);
+            if (!isAuthorized.Succeeded)
+            {
+                return Forbid();
             }
 
             _context.LeagueDriver.Remove(leagueDriver);
